login control before action

+++ api/tasks/index.php
@@ -20,41 +20,61 @@ if(gettype($conn) == 'object'){
 
     $task = new Task($conn);
 
-    if($method == 'POST') {
-        if(isset($data->user_id) && isset($data->title) && isset($data->date_time)){
-            $task->user_id = $data->user_id;
-            $task->title = $data->title;
-            $task->date_time = $data->date_time;
+    session_start();
 
-            $response = $task->add();
-        } else {
-            $response = array('message' => 'failed');
+    if(isset($_SESSION['logged'])){
+
+        if($method == 'POST') {
+            if(isset($_SESSION['id']) && isset($data->title) && isset($data->date_time)){
+                $task->user_id = $_SESSION['id'];
+                $task->title = $data->title;
+                $task->date_time = $data->date_time;
+    
+                $response = $task->add();
+            } else {
+                $response = array('message' => 'failed');
+            }
+        } else if($method == 'GET') {
+            $task->user_id = $_SESSION['id'];
+            $task->date_time = $_GET['date_time'];
+    
+            $response = $task->getTasksFromUser();
+        } else if($method == 'PUT') {
+            if(isset($data->id)){
+                $task->id = $data->id;
+    
+                $userID = $task->getTaskById();
+    
+                if($userID['user_id'] == $_SESSION['id']){
+                    if(isset($data->title)){$task->title = $data->title;}else{$task->title = null;}
+                    if(isset($data->date_time)){$task->date_time = $data->date_time;}else{$task->date_time = null;}
+                    if(isset($data->done)){$task->done = $data->done;}else{$task->done = null;}
+                    $response = $task->edit();
+                } else {
+                    $response = array('message' => 'Not logged.');
+                }
+            } else {
+                $response = array('message' => 'failed');
+            }
+        } else if($method == 'DELETE') {
+            if(isset($data->id)){
+                $task->id = $data->id;
+    
+                $userID = $task->getTaskById();
+                if($userID['user_id'] == $_SESSION['id']){
+                    $response = $task->delete();
+                } else {
+                    $response = array('message' => 'Not logged.');
+                }
+            } else {
+                $response = array('message' => 'failed');
+            }
         }
-    } else if($method == 'GET') {
-        $task->user_id = $_GET['user_id'];
-        $task->date_time = $_GET['date_time'];
 
-        $response = $task->getTasksFromUser();
-    } else if($method == 'PUT') {
-        if(isset($data->id)){
-            $task->id = $data->id;
-            if(isset($data->title)){$task->title = $data->title;}else{$task->title = null;}
-            if(isset($data->date_time)){$task->date_time = $data->date_time;}else{$task->date_time = null;}
-            if(isset($data->done)){$task->done = $data->done;}else{$task->done = null;}
-
-            $response = $task->edit();
-        } else {
-            $response = array('message' => 'failed');
-        }
-    } else if($method == 'DELETE') {
-        if(isset($data->id)){
-            $task->id = $data->id;
-
-            $response = $task->delete();
-        } else {
-            $response = array('message' => 'failed');
-        }
+    } else {
+        $response = array('message' => 'Not logged.');
     }
+
 } else {
     $response = $conn;
 }
